# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fall Prevention Plan API
 * OpenAPI spec version: 0.1.0
 */
export type ImageDecisionInputDecision = typeof ImageDecisionInputDecision[keyof typeof ImageDecisionInputDecision];


export const ImageDecisionInputDecision = {
  approve: 'approve',
  reject: 'reject',
} as const;

export interface ImageDecisionInput {
  decision: ImageDecisionInputDecision;
  /** @nullable */
  file?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  reviewer?: string | null;
}

export type ImageDecisionDecision = typeof ImageDecisionDecision[keyof typeof ImageDecisionDecision];


export const ImageDecisionDecision = {
  approve: 'approve',
  reject: 'reject',
} as const;

export interface ImageDecision {
  slotId: string;
  decision: ImageDecisionDecision;
  /** @nullable */
  file?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  reviewer?: string | null;
  updatedAt: string;
}

export interface HealthStatus {
  status: string;
}

/**
 * Access tier for the current user
 */
export type Tier = typeof Tier[keyof typeof Tier];


export const Tier = {
  guest: 'guest',
  one_time: 'one_time',
  subscription: 'subscription',
  concierge: 'concierge',
  admin: 'admin',
} as const;

export interface Me {
  signedIn: boolean;
  tier: Tier;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  name?: string | null;
  /** @nullable */
  joinedAt?: string | null;
}

export interface AssessmentOption {
  value: string;
  label: string;
  /** Points contributed to risk score when selected */
  score: number;
}

export interface AssessmentQuestion {
  id: string;
  /** STEADI category, e.g. history, mobility, vision, meds, home */
  category?: string;
  prompt: string;
  /** @nullable */
  helpText?: string | null;
  options: AssessmentOption[];
}

export type AssessmentInputAnswersItem = {
  questionId: string;
  value: string;
};

export interface AssessmentInput {
  answers: AssessmentInputAnswersItem[];
  /** @nullable */
  name?: string | null;
  /** @nullable */
  email?: string | null;
}

export type AssessmentResultLevel = typeof AssessmentResultLevel[keyof typeof AssessmentResultLevel];


export const AssessmentResultLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export interface AssessmentResult {
  score: number;
  level: AssessmentResultLevel;
  headline: string;
  summary?: string;
  recommendations: string[];
  completedAt: string;
}

export type PlanSection = typeof PlanSection[keyof typeof PlanSection];


export const PlanSection = {
  welcome: 'welcome',
  intro: 'intro',
  overview: 'overview',
  assessment: 'assessment',
  ten_point: 'ten_point',
  fall_response: 'fall_response',
  appendix_a: 'appendix_a',
  appendix_b: 'appendix_b',
  summary: 'summary',
} as const;

export interface ModuleSummary {
  slug: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  order: number;
  planSection: PlanSection;
  /** @nullable */
  durationMin?: number | null;
  locked: boolean;
  comingSoon?: boolean;
}

export interface Module {
  slug: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  order: number;
  planSection: PlanSection;
  /** @nullable */
  durationMin?: number | null;
  locked: boolean;
  comingSoon: boolean;
  /** @nullable */
  videoEmbedUrl?: string | null;
  /**
     * Markdown body, only present when not locked
     * @nullable
     */
  body?: string | null;
  keyPoints?: string[];
  printable?: boolean;
}

export type LiveSessionKind = typeof LiveSessionKind[keyof typeof LiveSessionKind];


export const LiveSessionKind = {
  class: 'class',
  qa: 'qa',
  one_on_one: 'one_on_one',
} as const;

export interface LiveSession {
  id: number;
  kind: LiveSessionKind;
  title: string;
  /** @nullable */
  description?: string | null;
  startsAt: string;
  durationMin: number;
  host: string;
  /**
     * Mocked join URL (stub)
     * @nullable
     */
  joinUrl?: string | null;
}

export type LibraryItemKind = typeof LibraryItemKind[keyof typeof LibraryItemKind];


export const LibraryItemKind = {
  recording: 'recording',
  article: 'article',
  interview: 'interview',
} as const;

export interface LibraryItem {
  id: number;
  title: string;
  kind: LibraryItemKind;
  /** @nullable */
  summary?: string | null;
  publishedAt: string;
  /** @nullable */
  durationMin?: number | null;
}

export interface Invoice {
  id: string;
  description: string;
  amountCents: number;
  paidAt: string;
}

export interface BillingSummary {
  tier: Tier;
  /** @nullable */
  priceLabel?: string | null;
  /** @nullable */
  nextChargeAt?: string | null;
  invoices: Invoice[];
}

export type CheckoutInputPlan = typeof CheckoutInputPlan[keyof typeof CheckoutInputPlan];


export const CheckoutInputPlan = {
  one_time: 'one_time',
  subscription: 'subscription',
  concierge: 'concierge',
} as const;

export interface CheckoutInput {
  plan: CheckoutInputPlan;
}

export interface ConciergeCheckIn {
  id: number;
  message: string;
  fromName: string;
  createdAt: string;
}

export interface ConciergeNote {
  id: number;
  body: string;
  authorName: string;
  createdAt: string;
}

export interface ConciergeDashboard {
  memberName: string;
  nextOutreachAt: string;
  latestCheckIn: ConciergeCheckIn;
  notes: ConciergeNote[];
}

/**
 * @nullable
 */
export type AdminMemberRiskLevel = typeof AdminMemberRiskLevel[keyof typeof AdminMemberRiskLevel] | null;


export const AdminMemberRiskLevel = {
  low: 'low',
  moderate: 'moderate',
  high: 'high',
} as const;

export interface AdminMember {
  id: number;
  name: string;
  email: string;
  tier: Tier;
  signupDate: string;
  lastLogin: string;
  /** @nullable */
  assessmentScore?: number | null;
  /** @nullable */
  riskLevel?: AdminMemberRiskLevel;
}

export interface AdminModule {
  slug: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  order: number;
  planSection: PlanSection;
  /** @nullable */
  durationMin?: number | null;
  /** @nullable */
  videoEmbedUrl?: string | null;
  /** @nullable */
  body?: string | null;
  keyPoints: string[];
  comingSoon: boolean;
  freeTier: boolean;
  printable: boolean;
}

export interface AdminModuleInput {
  slug: string;
  title: string;
  /** @nullable */
  subtitle?: string | null;
  order: number;
  planSection: PlanSection;
  /** @nullable */
  durationMin?: number | null;
  /** @nullable */
  videoEmbedUrl?: string | null;
  /** @nullable */
  body?: string | null;
  keyPoints?: string[];
  comingSoon?: boolean;
  freeTier?: boolean;
  printable?: boolean;
}

export interface AdminStats {
  totalMembers: number;
  subscriptionMembers: number;
  conciergeMembers: number;
  oneTimeMembers: number;
  assessmentsCompleted: number;
  mrrCents: number;
}

export type DeleteAdminModule200 = {
  ok: boolean;
};