# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fall Prevention Plan API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /assessment/submit:
    post:
      operationId: submitAssessment
      tags: [assessment]
      summary: Submit answers and get a scored fall-risk profile (public)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AssessmentInput"
      responses:
        "200":
          description: Scored result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AssessmentResult"
...
  /modules/{slug}:
    get:
      operationId: getModule
      tags: [modules]
      summary: Get a single module (full body if unlocked)
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Module"
...
  /billing/checkout:
    post:
      operationId: mockCheckout
      tags: [billing]
      summary: Mocked Stripe checkout — flips the signed-in user's tier
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CheckoutInput"
      responses:
        "200":
          description: Updated billing summary after mock purchase
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BillingSummary"
...
  /admin/modules:
    get:
      operationId: listAdminModules
      tags: [admin]
      summary: List every module with full editable fields (admin only)
      responses:
        "200":
          description: Modules
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AdminModule"
    post:
      operationId: createAdminModule
      tags: [admin]
      summary: Create a new module
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminModuleInput"
      responses:
        "200":
          description: Created module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminModule"
  /admin/modules/{slug}:
    get:
      operationId: getAdminModule
      tags: [admin]
      summary: Get a single module with full editable fields
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminModule"
    put:
      operationId: updateAdminModule
      tags: [admin]
      summary: Update an existing module
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminModuleInput"
      responses:
        "200":
          description: Updated module
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminModule"
    delete:
      operationId: deleteAdminModule
      tags: [admin]
      summary: Delete a module
      parameters:
        - in: path
          name: slug
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Deletion result
          content:
            application/json:
              schema:
                type: object
                required: [ok]
                properties:
                  ok: { type: boolean }
...
  /admin/image-decisions/{slotId}:
    put:
      operationId: upsertImageDecision
      tags: [admin]
      summary: Record or update the decision for one image slot
      parameters:
        - name: slotId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImageDecisionInput"
      responses:
        "200":
          description: Saved decision
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImageDecision"
components:
  schemas:
    ImageDecisionInput:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [approve, reject]
        file: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        reviewer: { type: ["string", "null"] }
    ImageDecision:
      type: object
      required: [slotId, decision, updatedAt]
      properties:
        slotId: { type: string }
        decision:
          type: string
          enum: [approve, reject]
        file: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        reviewer: { type: ["string", "null"] }
        updatedAt: { type: string, format: date-time }
...
    Tier:
      type: string
      enum: [guest, one_time, subscription, concierge, admin]
      description: Access tier for the current user
...
    AssessmentInput:
      type: object
      required: [answers]
      properties:
        answers:
          type: array
          items:
            type: object
            required: [questionId, value]
            properties:
              questionId: { type: string }
              value: { type: string }
        name: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
    AssessmentResult:
      type: object
      required: [score, level, headline, recommendations, completedAt]
      properties:
        score: { type: integer }
        level:
          type: string
          enum: [low, moderate, high]
        headline: { type: string }
        summary: { type: string }
        recommendations:
          type: array
          items: { type: string }
        completedAt: { type: string, format: date-time }
    PlanSection:
      type: string
      enum: [welcome, intro, overview, assessment, ten_point, fall_response, appendix_a, appendix_b, summary]
...
    Module:
      type: object
      required: [slug, title, order, planSection, locked, comingSoon]
      properties:
        slug: { type: string }
        title: { type: string }
        subtitle: { type: ["string", "null"] }
        order: { type: integer }
        planSection: { $ref: "#/components/schemas/PlanSection" }
        durationMin: { type: ["integer", "null"] }
        locked: { type: boolean }
        comingSoon: { type: boolean }
        videoEmbedUrl: { type: ["string", "null"] }
        body: { type: ["string", "null"], description: "Markdown body, only present when not locked" }
        keyPoints:
          type: array
          items: { type: string }
        printable: { type: boolean }
...
    Invoice:
      type: object
      required: [id, description, amountCents, paidAt]
      properties:
        id: { type: string }
        description: { type: string }
        amountCents: { type: integer }
        paidAt: { type: string, format: date-time }
    BillingSummary:
      type: object
      required: [tier, invoices]
      properties:
        tier: { $ref: "#/components/schemas/Tier" }
        priceLabel: { type: ["string", "null"] }
        nextChargeAt: { type: ["string", "null"], format: date-time }
        invoices:
          type: array
          items: { $ref: "#/components/schemas/Invoice" }
    CheckoutInput:
      type: object
      required: [plan]
      properties:
        plan:
          type: string
          enum: [one_time, subscription, concierge]
...
    AdminModule:
      type: object
      required: [slug, title, order, planSection, comingSoon, freeTier, printable, keyPoints]
      properties:
        slug: { type: string }
        title: { type: string }
        subtitle: { type: ["string", "null"] }
        order: { type: integer }
        planSection: { $ref: "#/components/schemas/PlanSection" }
        durationMin: { type: ["integer", "null"] }
        videoEmbedUrl: { type: ["string", "null"] }
        body: { type: ["string", "null"] }
        keyPoints:
          type: array
          items: { type: string }
        comingSoon: { type: boolean }
        freeTier: { type: boolean }
        printable: { type: boolean }
    AdminModuleInput:
      type: object
      required: [slug, title, order, planSection]
      properties:
        slug: { type: string }
        title: { type: string }
        subtitle: { type: ["string", "null"] }
        order: { type: integer }
        planSection: { $ref: "#/components/schemas/PlanSection" }
        durationMin: { type: ["integer", "null"] }
        videoEmbedUrl: { type: ["string", "null"] }
        body: { type: ["string", "null"] }
        keyPoints:
          type: array
          items: { type: string }
        comingSoon: { type: boolean }
        freeTier: { type: boolean }
        printable: { type: boolean }
...