# lib/api-zod/src/generated/types/imageDecision.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Fall Prevention Plan API
 * OpenAPI spec version: 0.1.0
 */
import type { ImageDecisionDecision } from './imageDecisionDecision';

export interface ImageDecision {
  slotId: string;
  decision: ImageDecisionDecision;
  /** @nullable */
  file?: string | null;
  /** @nullable */
  notes?: string | null;
  /** @nullable */
  reviewer?: string | null;
  updatedAt: Date;
}